% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agennls.R
\name{agennls}
\alias{agennls}
\title{automatic generate a fromula and start values for unknown parameters for a mixture of adaptive or generalized Gaussian distribution}
\usage{
agennls(A, u, sig, r)
}
\arguments{
\item{A}{is the amplitude of waveform.}

\item{u}{is the peak location of waveform.}

\item{sig}{is the echo width of waveform.}

\item{r}{is the rate parameter of adpative or generalized Gaussian distribution. For Gaussian distributionm, r is fixed to 2.}
}
\value{
return a formula suitable for different number of waveform componments with adaptive Gaussian distribution.
}
\description{
The function allows you to prepare the formula and start values for adaptive or generalized Gaussian decomposition.
}
\examples{

A<-c(76,56,80);u<-c(29,40,67);sig<-c(4,3,2.5); r<-c(2,2,2) ##these four should have the same length
fg<-agennls(A,u,sig,r)
##input formula for Gaussian decomposition
fgf<-fg$formula
###start values
fgs<-fg$start
}
