% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decom.adaptive.R
\name{decom.adaptive}
\alias{decom.adaptive}
\title{decom.adaptive}
\usage{
decom.adaptive(x, smooth = TRUE, thres = 0.22, width = 3)
}
\arguments{
\item{x}{is a waveform with a index at the begining and followed with intensities.}

\item{smooth}{is tell whether you want to smooth the waveform to remove some obvious outliers. Default is TRUE.}

\item{thres}{is to determine if the detected peak is the real peak whose intensity should be higher than threshold*maximum intensity. Default is 0.22.}

\item{width}{width of moving window.Default is 3, must be odd integer between 1 and n.This parameter ONLY work when the smooth is TRUE.}
}
\value{
A list contains estimates of A, u, sig and ri (rate parameter in adaotive Gaussian function) after decomposition.
}
\description{
The function allows you to eatimate parameters charcterizing waveforms and to pave the way for generating waveform-based point cloud.
}
\examples{
##import return waveform data
data(return)
lr<-nrow(return)
ind<-c(1:lr)
return<-data.frame(ind,return)
x<-return[1,] ###must be a dataset including intensity with index at the beginning.
r1<-decom(x)

r2<-decom.adaptive(x)

# for the whole dataset
dr3<-apply(return,1,decom.adaptive)
dd<-return[10:100,]
dr4<-apply(dd,1,decom.adaptive)

####to collect all data
## waveform is correctly decomposed with index,some are not correct index by NA
rfit3<-do.call("rbind",lapply(dr3,"[[",1))

###the original results, which can help to get more detailed results.
ga3<-do.call("rbind",lapply(dr3,"[[",2))

###useful decompostion results for next step or geolocation transformation.
pa3<-do.call("rbind",lapply(dr3,"[[",3))

colnames(pa3)<-c("index","pi","t","sd","ri","pise","tse","sdse","rise")

####delete some wrong ones
rid<-rfit3[!is.na(rfit3),]
wid<-setdiff(c(1:lr),rid)  ###index of waveforms needs to be reprocessed
#Generally the adaptive Guassian will give less NA results comparing to Gaussian function (decom).
rpars<-pa3[!is.na(pa3[,1]),]    ###useful decomposition parameters
}
\references{
Tan Zhou, and Sorin C. Popescu, 2017. Bayesian decomposition of full waveform LiDAR data with uncertainty analysis.
  Remote Sensing of Environment 200 (2017): 43-62.
}
