% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decom.R
\name{decom}
\alias{decom}
\title{decom: decompose waveform with different functions including Gaussian, adpative Gaussian and Weibull functions.}
\usage{
decom(x, smooth = "TRUE", thres = 0.22, width = 3, fun = "G")
}
\arguments{
\item{x}{is a waveform with a index at the begining and followed with intensities.}

\item{smooth}{is tell whether you want to smooth the waveform to remove some obvious outliers. Default is TRUE.}

\item{thres}{is to determine if the detected peak is the real peak whose intensity should be higher than threshold*maximum intensity. Default is 0.22.}

\item{width}{width of moving window.Default is 3, must be odd integer between 1 and n.This parameter ONLY work when the smooth is TRUE.}

\item{fun}{determine which method will be used for waveform decompsotion.Default is to use Gaussian (G) function. Other two functions are adpative Gaussian (AG) and Weibull (W) functions.}
}
\value{
A list contains estimates of A, u, sig after decomposition.
}
\description{
The function allows you to eatimate parameters charcterizing waveforms and to pave the way for generating waveform-based point cloud.
}
\examples{

##import return waveform data
data(return)
lr<-nrow(return)
ind<-c(1:lr)
return<-data.frame(ind,return)
x<-return[182,] ###must be a dataset including intensity with index at the beginning.
r1<-decom(x)
r2<-decom(x,smooth="TRUE",width=3) ###you can assign different smooth width for the data

for the whole dataset
dr<-apply(return,1,decom)

####to collect all data
rfit<-do.call("rbind",lapply(dr,"[[",1)) ## waveform is correctly decomposed with index,some are not correct index by NA
ga<-do.call("rbind",lapply(dr,"[[",2))   ###the original results, which can help to get more detailed results.
pa<-do.call("rbind",lapply(dr,"[[",3))   ###useful decompostion results for next step or geolocation transformation.

colnames(pa)<-c("index","pi","t","sd","pise","tse","sdse")

####delete some wrong ones
rid<-rfit[!is.na(rfit),]
wid<-setdiff(c(1:lr),rid)  ###index of waveforms needs to be reprocessed
rpars<-pa[!is.na(pa[,1]),]    ###useful decomposition parameters
}
