% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decom_result}
\alias{decom_result}
\title{Results of decomposition using Gaussian model}
\format{
A data frame with 1000 rows and 7 variables
\describe{
  \item{index}{index of waveform, which can tell which rows or results belongs to specific waveform}
  \item{A}{the amplitude of one waveform componment (A)}
  \item{u}{the time location corresponds to the amplitude for one waveform componment}
  \item{sigma}{the echo width of one waveform componment}
  \item{A_std}{the standard error of A, which can be used for uncertainty analysis}
  \item{u_std}{the standard error of u, which can be used for uncertainty analysis}
  \item{sig_std}{the standard error of sigma, which can be used for uncertainty analysis}
}
}
\usage{
decom_result
}
\description{
This dataset contained the result from the Gaussian decompositon.
}
\keyword{datasets}
