% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gennls.R
\name{gennls}
\alias{gennls}
\title{automatic generate a fromula and start values for unknown parameters for a mixture of Gaussian distribution}
\usage{
gennls(A, u, sig)
}
\arguments{
\item{A}{is the amplitude of waveform.}

\item{u}{is the peak location of waveform.}

\item{sig}{is the echo width of waveform.}
}
\value{
return a formula suitable for different number of waveform componments with Gaussian distribution.The number of componments
        needs to be determined by the users. Generally you need to use peakfind function to roughly estimate the number of waveform componments
}
\description{
The function allows you to prepare the formula and start values for Gaussian decomposition
}
\examples{

A<-c(76,56,80);u<-c(29,40,67);sig<-c(4,3,2.5) ##these three should have the same length
fg<-gennls(A,u,sig)
##input formula for Gaussian decomposition
fgf<-fg$formula
###start values
fgs<-fg$start
}
