% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo}
\alias{geo}
\title{The reference geolocation of the return waveforms.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 500 rows and 17 columns.}
\usage{
geo
}
\description{
This dataset should have the same number of rows of the return waveforms (return).
It is generally coming with waveform data and provided by the data provider. The number of rows should be
the same as the return waveform or pulse (return).
}
\keyword{datasets}
