% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotransform.R
\name{geotransform}
\alias{geotransform}
\title{geotransform}
\usage{
geotransform(decomp, geo, source = "decomposition")
}
\arguments{
\item{decomp}{the object from the decomposition or after deconvolution and decomposition.}

\item{geo}{the reference geolocation that is generally coming with waveform data and provided by the data provider.}

\item{source}{is determined by input data. If estimated parameters are from dcomposition, source = "decomposition". Otherwise will be deconvolution and decomposition.
Default is decomposition.}
}
\description{
The function allows you to convert parameters from decomposition to point cloud by using georeference data (generally it should come with waveform data).
  Detailed description of how to calculate can refer to Zhou, T., Popescu, S.C., Krause, K., Sheridan, R.D., Putman, E., 2017.
  Gold - a noveldeconvolution algorithm with optimization for waveform LiDAR processing.ISPRS Journal of Photogrammetry and Remote Sensing 129 (2017): 131-150.
  For the direct decomposition method, three kinds of position are calculated: leading edge, peak and trail edge. For the deconvolution and decomposition method,
  only the peak position is used to calculate the position. In additional, the uncertianty of the point cloud was provided based on detected peaks'
  95% confidence interval using deterministic method for waveform decomposition. This fucntion is suitable for NEON waveform lidar. For other kinds of dataset,
  you need to preprocess data to the same format or have the same required information or datsets.
}
\examples{

data(geo)
data(decom_result)
data(decon_result)
##used part of data to show the results
decomp<-decom_result[1:80,]
geo<-geo[1:80]
deconp<-decon_result[1:80]

##the follwoing steps are reuired to conduct the geotransformation,
##we need assign exactly same column names
geoindex=c(1:9,16)
colnames(geo)[geoindex]<-c("index","orix","oriy","oriz","dx","dy","dz","outref","refbin","outpeak")
####for decomposition results

geor<-geotransform(decomp,geo)

########for deconvolution and decomposition
decongeo<-geotransform(decomp=deconp,geo,source="deconvolution")

}
