% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro.R
\name{gro}
\alias{gro}
\title{gro: find the ground location in a waveform (assume the last echo or peak corresponding to the ground).}
\usage{
gro(y, smooth = "TRUE", thres = 0.2, width = 3)
}
\arguments{
\item{smooth}{is tell whether you want to smooth the waveform to reduce the effect of some obvious noise. Default is TRUE.}

\item{thres}{is to determine if the detected peak is the real peak whose intensity should be higher than threshold*maximum intensity. Default is 0.22.}

\item{width}{the width of moving window for smoothing.Default is 3, must be integer between 1 and n.This parameter ONLY work when the smooth is TRUE.}

\item{x}{is the waveform intensities. If you have other information, you should delete these intensites before you run this function .}
}
\value{
return the index of possible ground position of waveform.
}
\description{
The function allows you to identify the possible ground location (index) in the waveform.
}
\examples{

data(return)
x<-return[125,]
#default
gro(x)

}
