% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imp}
\alias{imp}
\title{prototype system impulse}
\format{A data frame with 100 rows and 1 column.
\describe{
 \item{V1}{The intensity of the system impulse}
}}
\usage{
imp
}
\description{
This data was generated from a hard ground target with a mirror angle close to nadir
and its corresponding outgoing waveform. Generally one study site only need one system impulse which is
mainly for calibration purpose.
}
\keyword{datasets}
