% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imp_out}
\alias{imp_out}
\title{Outgoing pulse corresponding to the system impulse}
\format{A vector with the length of 100. The zero padding was used to make sure every waveforms has the same length. In this case, 0 means no recorded intensity.

\describe{
  \item{outgoing pulse}{}
}}
\usage{
imp_out
}
\description{
This dataset is the corresponding outgoing pulse of the system impulse
}
\keyword{datasets}
