% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral.R
\name{integral}
\alias{integral}
\title{integral: calculate the integral or cumulative enery from vegetation and ground parts of the waveform.}
\usage{
integral(y, smooth = "TRUE", thres = 0.2, width = 3, tr = 1, dis = 20)
}
\arguments{
\item{smooth}{is tell whether you want to smooth the waveform to remove some obvious outliers. Default is TRUE.}

\item{thres}{is to determine if the detected peak is the real peak whose intensity should be higher than threshold*maximum intensity. Default is 0.2.}

\item{width}{width of moving window.Default is 3, must be integer between 1 and n.This parameter ONLY work when the smooth is TRUE.}

\item{tr}{the temporal resolution of waveform.Default is 1 ns. Must be integer from 1 to n.}

\item{dis}{the distance from last echo (ground) which determine the boundary between ground and vegetation. Default is 20 ns which equals to 3 m.
This means the ground part signals are from identified ground location to 20 ns above and the signals of vegetation are from waveform begining to 3 m above ground.}

\item{x}{is the waveform intensities. If you have other information, you should delete these intensites before you run this function .}
}
\value{
return the integral of waveform intensity for the ground, vegetation parts, the whole waveform above ground and the integral ration betwwen vegetation and the whole waveform .
}
\description{
The function allows you to calculate the integral of vegetation part, ground part, sum of them and ration between vegation integral and total integral.
}
\examples{

data(return)
x<-return[1,]
#if we kept everything as default, before you use this function, you need to know the temporal resolution of waveform, default is 1 ns.
##for one peak, it generaly not make too much sense since generally only ground was present in this case.
r1<-integral(x)

#if you didn't want to smooth,
r2<-integral(x,smooth="FALSE")

##you also can define the boundary between vegetation and ground by assign adjusting dis
#if we assign 15, it means vegetation starts at 15*1*0.15 m above the last echo.

r3<-integral(x,dis=15)
# when it comes to the waveform with several peaks
xx<-return[182,]

rr1<-integral(xx)

}
