% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxa.R
\name{maxa}
\alias{maxa}
\title{maxa: find the maximum amplitude of waveform above the ground (not the whole waveform).}
\usage{
maxa(y, smooth = "TRUE", thres = 0.2, width = 3)
}
\arguments{
\item{smooth}{is tell whether you want to smooth the waveform to remove some obvious outliers. Default is TRUE.}

\item{thres}{is to determine if the detected peak is the real peak whose intensity should be higher than threshold*maximum intensity. Default is 0.2.}

\item{width}{width of moving window.Default is 3, must be integer between 1 and n.This parameter ONLY work when the smooth is TRUE.}

\item{x}{is the waveform intensities. If you have other information, you should delete these intensites before you run this function .}
}
\value{
return the largest amplitude above the ground for the waveform with more than two peaks and ground amplitude for the waveform with one peak.
}
\description{
The function allows you to identify maximum intensity which can be used to calculate the height from waveform ending or waveform begining.
}
\examples{

data(return)
x<-return[1,]
#default
maxa(x) ###for one peak waveform, it didn't make too much sense

x1<-return[182,]
rx1<-maxa(x1)
}
