% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.height.R
\name{med.height}
\alias{med.height}
\title{med.height: calculate the position and height of half energy above ground (not waveform ending).}
\usage{
med.height(x, smooth = "TRUE", thres = 0.2, width = 3, tr = 1)
}
\arguments{
\item{x}{is the waveform intensities. If you have other information, you should delete these intensites before you run this function .}

\item{smooth}{is tell whether you want to smooth the waveform to remove some obvious outliers. Default is TRUE.}

\item{thres}{is to determine if the detected peak is the real peak whose intensity should be higher than threshold*maximum intensity. Default is 0.22.}

\item{width}{width of moving window.Default is 3, must be integer between 1 and n.This parameter ONLY work when the smooth is TRUE.}

\item{tr}{the temporal resolution of waveform.Default is 1 ns. Must be integer from 1 to n.}
}
\value{
return the height of half total energy above the ground and proportion of the first half.
}
\description{
The function allows you to identify the half total energy position which can be used to calculate the height from the ground.
}
\examples{

data(return)
x<-return[1,]
#default
med.height(x)  ###for one peak waveform, it generally not make too much sense
xx<-return[182,]
med.height(xx)  ##this can help to charcterize the waveform
}
