% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outg}
\alias{outg}
\title{500 corresponding outgoing pulses.}
\format{A data frame with 500 rows and 100 variables:}
\usage{
outg
}
\description{
This dataset should have the same number of rows as the return pulses, which is the outgoing
pulse for each return waveform pulse.The zero padding was used to make sure every waveforms has the same length. In this case, 0 means no recorded intensity.
}
\keyword{datasets}
