% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{peaks}
\alias{peaks}
\title{peak finding function}
\usage{
peaks(series, span = 3)
}
\arguments{
\item{series}{is the input a numeric vector.}

\item{span}{is the length or interval of peak finding cell, default is 3.}
}
\value{
return a boolean type data corresponding to the numeric vector.
}
\description{
This function allows you to find where is the peak and return a list of TRUE and FALSE
   TRUE represents this location is the peak.
}
\examples{
w1 is numeric values represnt an waveform.
peaks(w1,3)
peaks(w2,5)
}
