% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile.R
\name{percentile}
\alias{percentile}
\title{percentile: calculate the position of energy percentile in the waveform (the total waveform).}
\usage{
percentile(x, quan = seq(0.5, 1, 0.1), normalization = "TRUE")
}
\arguments{
\item{x}{is the waveform intensities. If you have other information, you should delete these intensites before you run this function .}

\item{quan}{is the quantile of energy you are interested in. Defalut is seq(0.5,1,0.1).}

\item{normalization}{is determine whether you want to normalized the waveform intensity or not. Here we used the minimum intensity for each waveform to conduct normalization.
Default is use normalization.}
}
\value{
return the index or position of corressponding quantile energy of waveform.
}
\description{
The function allows you to calculate the position of energy percentile based on the given quantile within a waveform.
}
\examples{

data(return)
x<-return[1,]
#default
percentile(x)

#change the quantile
qr<-seq(0.45,0.99,0.03)
re1<-percentile(x,quan=qr)
###normalization affects your results
re2<-percentile(x,quan=qr,normalization = "FALSE")
##after you get the index, you can convert them to the height based on temporal resolution or georeference information of this waveform.
}
