% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{return}
\alias{return}
\title{500 sample LiDAR waveforms from Harvard Forest provided by National Ecological Observatory Networks (NEON).
More details can be found in	Tan Zhou*, Sorin C. Popescu, Keith Krause, Ryan D. Sheridan, and Eric Putman, 2017.
Gold-A novel deconvolution algorithm with optimization for waveform LiDAR processing. ISPRS Journal of Photogrammetry and Remote Sensing 129 (2017): 131-150. https://doi.org/10.1016/j.isprsjprs.2017.04.021}
\format{A data frame with 500 rows and 208 variables:}
\usage{
return
}
\description{
A dataset containing the return waveforms with intensity along the path. The zero padding was used
to make sure every waveforms has the same length. In this case, 0 means no recorded intensity.
}
\keyword{datasets}
