% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_hf}
\alias{shp_hf}
\title{A boundary shapefile, polygon}
\format{A shapefile containing a small boudary of the Harvard Forest.}
\usage{
shp_hf
}
\description{
This data is manily to test the fucntion waveformclip.
}
\keyword{datasets}
