% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decom.weibull.R
\name{decom.weibull}
\alias{decom.weibull}
\title{decom.weibull}
\usage{
decom.weibull(x, smooth = TRUE, thres = 0.22, width = 3)
}
\arguments{
\item{x}{is a waveform with a index at the begining and followed with intensities.}

\item{smooth}{is tell whether you want to smooth the waveform to remove some obvious outliers. Default is TRUE.}

\item{thres}{is to determine if the detected peak is the real peak whose intensity should be higher than threshold*maximum intensity. Default is 0.22.}

\item{width}{width of moving window.Default is 3, must be odd integer between 1 and n.This parameter ONLY work when the smooth is TRUE.}
}
\value{
A list contains estimates of A (amplitude parameter), u (location parameter), sigma (scale parameter) and k (shape parameter) after decomposition.
these parameters are different from the adaptive Gaussian and Gaussian decomposition.
}
\description{
The function allows you to fit waveforms with the Weibull function and get parameters estimated.
}
\examples{

##import return waveform data
data(return)
lr<-nrow(return)
ind<-c(1:lr)
return<-data.frame(ind,return)
x<-return[1,] ###must be a dataset including intensity with index at the beginning.
r1<-decom.weibull(x)

r2<-decom.weibull(return[2,])

# for the whole dataset
dr3<-apply(return,1,decom.weibull)
dd<-return[10:100,]
dr4<-apply(dd,1,decom.weibull)

####to collect all data
rfit3<-do.call("rbind",lapply(dr3,"[[",1)) ## waveform is correctly decomposed with index,some are not correct index by NA
ga3<-do.call("rbind",lapply(dr3,"[[",2))   ###the original results, which can help to get more detailed results.
pa3<-do.call("rbind",lapply(dr3,"[[",3))   ###useful decompostion results for next step or geolocation transformation.

colnames(pa3)<-c("index","pi","t","sd","ri","pise","tse","sdse","rise")

####delete some wrong ones
rid<-rfit3[!is.na(rfit3),]
wid<-setdiff(c(1:lr),rid)  ###index of waveforms needs to be reprocessed
rpars<-pa1[!is.na(pa1[,1]),]    ###useful decomposition parameters
Generally using adaptive Guassian will give less NA results comparing to Gaussian function (decom).
}
