% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolution}
\alias{deconvolution}
\title{deconvolution: deconvolove waveform to obtain effective target positions.}
\usage{
deconvolution(re, out, imp, method = c("Gold"), np = 2,
  normalization = "TRUE", para1 = c(30, 2, 1.8, 30, 2, 2), para2 = c(40,
  3, 1.8, 40, 3, 2))
}
\arguments{
\item{re}{is the return waveform.}

\item{out}{is the outgoing waveform.}

\item{imp}{is the impulse reponse.}

\item{method}{two methods including Gold or Richardson-Lucy (RL) algorithms. method=c("Gold","RL"). Default is method=c("Gold").}

\item{np}{a threshold value which determines use para1 (smaller iterations and repetitions in the deconvolution) or 
or para2 (larger iterations and repetitionsin the deconvolution). Default is 2.}

\item{normalization}{is determine whether you want to normalized the waveform intensity or not. Here we used the minimum intensity to conduct normalization. Default is use normalization.}

\item{para1}{is the parameters including iterations, repetitions and boost for deconvolution when the number of waveform componments is less than np.}

\item{para2}{is the parameters including iterations, repetitions and boost for deconvolution when the number of waveform componments is more than np.
Generally when the waveform become more complicated, these parameters should be larger than corresponding para1.
iterations: number of iterations (parameter L in the Gold deconvolution algorithm) between boosting operations;
repetitions number of repetitions of boosting operations. It must be greater or equal to one.So the total number of iterations is repetitions*iterations
boosting coefficient/exponent. Applies only if repetitions is greater than one. Recommended range [1..2].}
}
\value{
The deconvovled waveform.
}
\description{
The function allows you to deconvolve retrun waveform lidar data by using the corresponding outgoing waveform and impulse response to
  to obtain effective target positions. Two algorithms were used including Gold or Richardson-Lucy algorithms. The detailed deconvolution
  infromation can be founded in package Peaks.
}
\examples{

data(return)
data(outg)
data(imp)

re<-return[1,]
out<-outg[1,]
imp<-imp

dr<-deconvolution(re,out,imp)
dr1<-deconvolution(re,out,imp,method="RL")
dr2<-deconvolution(re,out,imp,method="RL",para1=c(20,2,1.8,20,2,2))

plot(dr,type="l")
lines(dr1,col="red")
lines(dr2,col="blue")
###some differences can be observed when you used different parameters.
In the real application, you need to find one suitable for your cases. 
}
