% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo}
\alias{geo}
\title{The reference geolocation of the return waveforms.}
\format{A data frame with 500 rows and 17 variables.
\describe{
 \item{index}{index to connect return waveform and geolocation}
 \item{\code{x}}{Easting of first return}
 \item{\code{y}}{Northing  of first return}
 \item{\code{z}}{Height of first return}
 \item{\code{dx}}{the position change per nanasecond at x direction}
 \item{\code{dy}}{the position change per nanasecond at y direction}
 \item{\code{dz}}{the position change per nanasecond at z direction}
 \item{\code{or}}{outgoing pulse reference bin location (leading edge 50th point of the outgoing pulse)}
 \item{\code{fr}}{first return reference bin location (leading edge 50th point of the first return)}
 \item{\code{V9}}{Easting of Return Bin0}
 \item{\code{V10}}{Northing of Return Bin0}
 \item{\code{V11}}{Height of Return Bin0}
 \item{\code{V12}}{the position change per nanasecond at x direction}
 \item{\code{V13}}{the position change per nanasecond at y direction}
 \item{\code{V14}}{the position change per nanasecond at z direction}
 \item{\code{V15}}{outgoing pulse peak bin location}
 \item{\code{V16}}{return bin0 location}
}}
\usage{
geo
}
\description{
This dataset should have the same number of rows of the return waveforms (return).
It is generally coming with waveform data and provided by the data provider. The number of rows should be
the same as the return waveform or pulse (return).
}
\keyword{datasets}
