% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelen.R
\name{wavelen}
\alias{wavelen}
\title{wavelen: calculate the length of waveform intensity}
\usage{
wavelen(x)
}
\arguments{
\item{x}{is the waveform intensities. If you have other information, you should delete these intensites before you run this function .}
}
\value{
return the the number of waveform intensities or waveform length.
}
\description{
The function allows you to measure the length of the waveform with intensity after excluding non-intensity parts.
Generally, we assumed the intensity of a waveform should be larger than 0.
}
\examples{

data(return)
y<-cbind(1:nrow(return),return)
x<-return[1,]
wavelen(x)
#for x has index and intensities, we should delete non-intensities first
yy<-y[,-1]
xx<-yy[1,]
wavelen(xx)
}
